refactoring the navbar component
